% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_selection_functions.R
\name{computeRho}
\alias{computeRho}
\title{Compute posterior predictive distribution of \code{rho.lambda}}
\usage{
computeRho(beta_path, XX, post_beta, post_trace_sigma_2)
}
\arguments{
\item{beta_path}{\code{M x P x L} array of regression coefficients \code{beta}
along the solution path of length \code{L}}

\item{XX}{\code{N x P} matrix of predictors}

\item{post_beta}{\code{Nsims x P x M} array of posterior draws of \code{beta}}

\item{post_trace_sigma_2}{\code{Nsims x 1} vector of posterior draws of the trace
of the (marginal) covariance (see below for details)}
}
\value{
A list containing \code{rho_lam0} and \code{rho_lam2},
corrosponding to rho2 for the full model and the sparsified model
(for each value of \code{lambda} in the solution path).
}
\description{
Using the (group) lasso solution path, compute the proportion of
variability explained by the sparsified models (relative to the full model)
for each MCMC simulation.
}
\note{
\code{post_trace_sigma_2} is the (posterior samples of)
the trace of the error covariance matrix jointly across subjects i=1,...,n
and observations j=1,...,m, after marginalizing out the random effects \code{gamma_ik}.
This is given by \code{nm x sigma_e^2} + \code{sum_ik sigma_gamma_ik^2},
where the second term is necessary only when random effects are included in the model
AND integrated over in the predictive distribution.
}
