% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_sampler.R
\name{dfosr_basis_ar}
\alias{dfosr_basis_ar}
\title{MCMC Sampling Algorithm for the Dynamic Function-on-Scalars Regression Model with AR(1) Errors
and a Known Basis Expansion}
\usage{
dfosr_basis_ar(Y, tau, X = NULL, use_dynamic_reg = TRUE, use_fpca = TRUE,
  use_shrinkage_priors = FALSE, nsave = 1000, nburn = 1000, nskip = 2,
  mcmc_params = list("beta", "fk", "alpha", "mu_k", "ar_phi"),
  use_obs_SV = FALSE, computeDIC = TRUE)
}
\arguments{
\item{Y}{the \code{T x m} data observation matrix, where \code{T} is the number of time points and \code{m} is the number of observation points (\code{NA}s allowed)}

\item{tau}{the \code{m x d} matrix of coordinates of observation points}

\item{X}{the \code{T x p} matrix of predictors; if NULL, only include an intercept}

\item{use_dynamic_reg}{logical; if TRUE, regression coefficients are dynamic
(with random walk models)}

\item{use_fpca}{logical; if TRUE, use functional principal components basis; otherwise use splines}

\item{use_shrinkage_priors}{logical; if TRUE, include shrinkage priors for the coefficients}

\item{nsave}{number of MCMC iterations to record}

\item{nburn}{number of MCMC iterations to discard (burin-in)}

\item{nskip}{number of MCMC iterations to skip between saving iterations,
i.e., save every (nskip + 1)th draw}

\item{mcmc_params}{named list of parameters for which we store the MCMC output;
must be one or more of
\itemize{
\item "beta" (factors)
\item "fk" (loading curves)
\item "alpha" (regression coefficients)
\item "mu_k" (intercept term for factor k)
\item "ar_phi" (AR coefficients for each k under AR(1) model)
\item "sigma_et" (observation error SD; possibly dynamic)
\item "Yhat" (fitted values)
\item "Ypred" (posterior predictive values)
}}

\item{use_obs_SV}{logical; when TRUE, include a stochastic volatility model
for the observation error variance}

\item{computeDIC}{logical; if TRUE, compute the deviance information criterion \code{DIC}
and the effective number of parameters \code{p_d}}
}
\value{
A named list of the \code{nsave} MCMC samples for the parameters named in \code{mcmc_params}
}
\description{
Runs the MCMC for the dynamic function-on-scalars regression model based on
a known basis expansion (splines or functional principal components).
Here, we assume the factor regression has AR(1) errors.
}
