% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_sampler.R
\name{dfosr_rw}
\alias{dfosr_rw}
\title{MCMC Sampling Algorithm for the Dynamic Function-on-Scalars Regression Model with Random Walk Errors}
\usage{
dfosr_rw(Y, tau, X = NULL, K = NULL, use_dynamic_reg = TRUE,
  nsave = 1000, nburn = 1000, nskip = 3, mcmc_params = list("beta",
  "fk", "alpha"), X_Tp1 = 1, use_obs_SV = FALSE,
  includeBasisInnovation = FALSE, Con_mat = NULL)
}
\arguments{
\item{Y}{the \code{T x m} data observation matrix, where \code{T} is the number of time points and \code{m} is the number of observation points (\code{NA}s allowed)}

\item{tau}{the \code{m x d} matrix of coordinates of observation points}

\item{X}{the \code{T x p} matrix of predictors; if NULL, only include an intercept}

\item{K}{the number of factors; if NULL, use SVD-based proportion of variability explained}

\item{use_dynamic_reg}{logical; if TRUE, regression coefficients are dynamic
(with random walk models), otherwise independent}

\item{nsave}{number of MCMC iterations to record}

\item{nburn}{number of MCMC iterations to discard (burin-in)}

\item{nskip}{number of MCMC iterations to skip between saving iterations,
i.e., save every (nskip + 1)th draw}

\item{mcmc_params}{named list of parameters for which we store the MCMC output;
must be one or more of
\itemize{
\item "beta" (dynamic factors)
\item "fk" (loading curves)
\item "alpha" (regression coefficients; possibly dynamic)
\item "mu_k" (intercept term for factor k)
\item "sigma_et" (observation error SD; possibly dynamic)
\item "Yhat" (fitted values)
\item "Ypred" (posterior predictive values)
\item "Yfore" (one-step forecast; includes the estimate and the distribution)
}}

\item{X_Tp1}{the \code{p x 1} matrix of predictors at the forecasting time point \code{T + 1}}

\item{use_obs_SV}{logical; when TRUE, include a stochastic volatility model
for the observation error variance}

\item{includeBasisInnovation}{logical; when TRUE, include an iid basis coefficient term for residual correlation
(i.e., the idiosyncratic error term for a factor model on the full basis matrix)}

\item{Con_mat}{a \code{Jc x m} matrix of constraints for the loading curves such that
\code{Con_mat*fk = 0} for each loading curve \code{fk}; default is NULL for no constraints.}
}
\value{
A named list of the \code{nsave} MCMC samples for the parameters named in \code{mcmc_params}
}
\description{
Runs the MCMC for the dynamic function-on-scalars regression model based on
an reduced-rank expansion. Here, we assume the factor regression has RW errors.
}
