% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_selection_functions.R
\name{dss_select}
\alias{dss_select}
\title{Group lasso for matrix regression}
\usage{
dss_select(Y, X, w = NULL)
}
\arguments{
\item{Y}{\code{N x M} matrix of observations}

\item{X}{\code{N x P} matrix of predictors}

\item{w}{\code{P x 1} vector of weights; if NULL, use sqrt(M)}
}
\value{
The solution path for \code{L} values of \code{lambda} in the form
of an array of dimension \code{M x P x L}
}
\description{
Given a \code{N x M} data matrix \code{Y}
and a \code{N x P} matrix of predictor \code{X},
estimate the model
\code{Y = XB + E}
for \code{P x M} regression coefficient matrix \code{B}.
The penalty applies a row-wise group lasso penalty, i.e.,
for row \code{p}, the \code{M} elements \code{X[p,]} are regularized toward zero.
}
\note{
The design matrix \code{X} may include an intercept, which will
be left unpenalized.
}
