% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_selection_functions.R
\name{dss_select_lasso}
\alias{dss_select_lasso}
\title{Lasso for matrix regression}
\usage{
dss_select_lasso(Y, X, W = NULL)
}
\arguments{
\item{Y}{\code{N x M} matrix of observations}

\item{X}{\code{N x P} matrix of predictors}

\item{W}{\code{P x M} matrix of weights}
}
\value{
The solution path for \code{L} values of \code{lambda} in the form
of an array of dimension \code{M x P x L}
}
\description{
Given a \code{N x M} data matrix \code{Y}
and a \code{N x P} matrix of predictor \code{X},
estimate the model
\code{Y = XB + E}
for \code{P x M} regression coefficient matrix \code{B}.
The penalty applies a lasso penalty to all elements of \code{B}.
}
