% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_sampler.R
\name{fdlm}
\alias{fdlm}
\title{MCMC Sampling Algorithm for the Functional Dynamic Linear Model}
\usage{
fdlm(Y, tau, K = NULL, factor_model = "RW", nsave = 1000,
  nburn = 1000, nskip = 3, mcmc_params = list("beta", "fk"),
  use_obs_SV = FALSE, X_Tp1 = 1, includeBasisInnovation = FALSE,
  Con_mat = NULL)
}
\arguments{
\item{Y}{the \code{T x m} data observation matrix, where \code{T} is the number of time points and \code{m} is the number of observation points (\code{NA}s allowed)}

\item{tau}{the \code{m x d} matrix of coordinates of observation points}

\item{K}{the number of factors; if NULL, use SVD-based proportion of variability explained}

\item{factor_model}{model for the (dynamic) factors;
must be one of
\itemize{
\item "IND" (independent errors)
\item "AR" (stationary autoregression of order 1)
\item "RW" (random walk model)
}}

\item{nsave}{number of MCMC iterations to record}

\item{nburn}{number of MCMC iterations to discard (burin-in)}

\item{nskip}{number of MCMC iterations to skip between saving iterations,
i.e., save every (nskip + 1)th draw}

\item{mcmc_params}{named list of parameters for which we store the MCMC output;
must be one or more of
\itemize{
\item "beta" (dynamic factors)
\item "fk" (loading curves)
\item "mu_k" (intercept)
\item "sigma_et" (observation error SD; possibly dynamic)
\item "Yhat" (fitted values)
\item "Ypred" (posterior predictive values)
\item "Yfore" (one-step forecast; includes the estimate and the distribution)
}}

\item{use_obs_SV}{logical; when TRUE, include a stochastic volatility model
for the observation error variance}

\item{X_Tp1}{the \code{p x 1} matrix of predictors at the forecasting time point \code{T + 1}}

\item{includeBasisInnovation}{logical; when TRUE, include an iid basis coefficient term for residual correlation
(i.e., the idiosyncratic error term for a factor model on the full basis matrix)}

\item{Con_mat}{a \code{Jc x m} matrix of constraints for the loading curves such that
\code{Con_mat*fk = 0} for each loading curve \code{fk}; default is NULL for no constraints.}
}
\value{
A named list of the \code{nsave} MCMC samples for the parameters named in \code{mcmc_params}
}
\description{
Runs the MCMC for the functional dynamic linear model (with no predictor variables).
Models for the (dynamic) factors include independent factors, an AR(1) model,
and a random walk model.
}
\note{
If \code{Tm} is large, then storing all posterior samples for \code{Yhat} or \code{Ypred}, which are \code{nsave x T x m},  may be inefficient
}
\examples{
# Simulate some data (w/o predictors):
sim_data = simulate_dfosr(T = 100, m = 20,
                          p_0 = 0, p_1 = 0)
Y = sim_data$Y; tau = sim_data$tau
T = nrow(Y); m = ncol(Y); # Dimensions

# Run the MCMC w/ K = 6:
out = fdlm(Y = Y, tau = tau, K = 6,
           factor_model = 'AR',
           mcmc_params = list("beta", "fk", "Yhat", "Ypred"))

# Plot the factors:
plot_factors(post_beta = out$beta)

# Plot the loading curves:
plot_flc(post_fk = out$fk, tau = tau)

# Plot a fitted value w/ posterior predictive credible intervals:
i = sample(1:T, 1); # Select a random time i
plot_fitted(y = Y[i,],
            mu = colMeans(out$Yhat)[i,],
            postY = out$Ypred[,i,],
            y_true = sim_data$Y_true[i,],
            t01 = tau)


}
