% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{fdlm_init}
\alias{fdlm_init}
\title{Initialize the reduced-rank functional data model}
\usage{
fdlm_init(Y, tau, K = NULL, use_pace = FALSE)
}
\arguments{
\item{Y}{the \code{T x m} data observation matrix, where \code{T} is the number of time points and \code{m} is the number of observation points (\code{NA}s allowed)}

\item{tau}{the \code{m x d} matrix of coordinates of observation points}

\item{K}{the number of factors; if \code{NULL}, select using proportion of variability explained (0.99)}

\item{use_pace}{logical; if TRUE, use the PACE procedure for FPCA (only for \code{d} = 1);
otherwise use splines}
}
\value{
a list containing
\itemize{
\item \code{Beta} the \code{T x K} matrix of factors
\item \code{Psi} the \code{J x K} matrix of loading curve basis coefficients,
where \code{J} is the number of spline basis functions
\item \code{splineInfo} a list with information about the spline basis
\item \code{Y0} the imputed data matrix
}
}
\description{
Compute initial values for the factors and loadings curves,
where the input matrix may have NAs.
}
\details{
The PACE procedure is useful for estimated a reduced-rank functional data model
in the case of sparsely-observed FD, i.e., many NAs in \code{Y}. However,
it is also only valid for univariate observation points
}
