% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{simulate_dfosr}
\alias{simulate_dfosr}
\title{Simulate a dynamic function-on-scalars regression model}
\usage{
simulate_dfosr(T = 200, m = 100, RSNR = 5, K_true = 4, p_0 = 10,
  p_1 = 5, use_dynamic_reg = TRUE, sparse_factors = TRUE, ar1 = 0,
  perc_missing = 0)
}
\arguments{
\item{T}{number of observed curves (i.e., number of time points)}

\item{m}{total number of observation points (i.e., points along the curve)}

\item{RSNR}{root signal-to-noise ratio}

\item{K_true}{rank of the model (i.e., number of basis functions used for the functional data simulations)}

\item{p_0}{number of true zero regression coefficients}

\item{p_1}{number of true nonzero regression coefficients}

\item{use_dynamic_reg}{logical; if TRUE, simulate dynamic regression coefficients; otherwise static}

\item{sparse_factors}{logical; if TRUE, then for each nonzero predictor j,
sample a subset of k=1:K_true factors to be nonzero#'}

\item{ar1}{AR(1) coefficient for time-correlated predictors}

\item{perc_missing}{percentage of missing data (between 0 and 1); default is zero}
}
\value{
a list containing the following:
\itemize{
\item \code{Y}: the simulated \code{T x m} functional data matrix
\item \code{X}: the simulated \code{T x p} design matrix
\item \code{tau}: the \code{m}-dimensional vector of observation points
\item \code{Y_true}: the true \code{T x m} functional data matrix (w/o noise)
\item \code{alpha_tilde_true} the true \code{T x p x m} array of regression coefficient functions
\item \code{alpha_arr_true} the true \code{T x p x K_true} array of (dynamic) regression coefficient factors
\item \code{Beta_true} the true \code{T x K_true} matrix of factors
\item \code{F_true} the true \code{m x K_true} matrix of basis (loading curve) functions
\item \code{sigma_true} the true observation error standard deviation
}
}
\description{
Simulate data from a dynamic function-on-scalars regression model, allowing for
autocorrelated errors and (possibly) dynamic regression coefficients random effects.
The predictors are contemporaneously independent but (possibly) autocorrelated.
}
\note{
The basis functions (or loading curves) are orthonormalized polynomials,
so large values of \code{K_true} are not recommended.
}
\examples{
# Example: simulate DFOSR
sim_data = simulate_dfosr()
Y = sim_data$Y; X = sim_data$X; tau = sim_data$tau

}
